Align skills, mission storage and packets, and logout with client evidence

+++ src/Rasa.DBL/Structures/World/NpcMissionEntry.cs
@@ -33,7 +33,7 @@ namespace Rasa.Structures.World
 
         [Column("category_id")]
         [Required]
-        public byte CategoryId { get; set; }
+        public uint CategoryId { get; set; }
 
         [Column("shareable")]
         [Required]
